create dirtywords
dirtywords is a curses replacement. It is build on a minimal set of
functions that can be implemented with a variety of backend, e.g. curses
or pygame.

It is relatively hard to install either curses or pygame. They can not
simply be installed with pip. So dirtywords offers an abstraction layer
that provides the best possible experience on the target machiene.